LayoutTests/fast/mediastream/image-capture-creation.html

<!DOCTYPE html>
<html>
<head>
    <meta charset='utf-8'>
    <title>ImageCapture creation</title>
    <script src='../../resources/testharness.js'></script>
    <script src='../../resources/testharnessreport.js'></script>
</head>
<body>
    <canvas id='canvas'></canvas>
    <video controls autoplay width=640 height=480 playsInline id='video'></video>
    <script>

        promise_test(async (test) => {
            const stream = await navigator.mediaDevices.getUserMedia({ audio: true });
            const [track] = stream.getAudioTracks();

            assert_true(track.enabled);
            assert_false(track.muted);
            assert_equals(track.readyState, 'live');

            assert_throws_dom('NotSupportedError', () => { new ImageCapture(track) })

        }, `Creating an ImageCapture from an audio track should throw "NotSupportedError"`);

    </script>
</body>
</html>
